backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.45.2",
    "fastapi>=0.115.8",
    "httpx>=0.28.1",
    "pydantic-settings>=2.7.1",
    "pydantic>=2.10.6",
    "pydantic-ai>=0.0.24",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
    "logfire[httpx]>=3.5.3",
]
